feat(test): agregar clase TestCalculadora para ejecutar metodos con retorno

+++ src/com/krakedev/Calculadora.java
@@ -2,9 +2,9 @@ package com.krakedev;
 
 public class Calculadora {
 	
-	public int sumar(int a, int b) {
+	public double sumar(int a, int b) {
 		
-		int resultado;
+		double resultado;
 		resultado = a + b;
 		
 		return resultado;
